+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added additional output options (cvs, toa5, feather, parquet, or pandas)
 * TOA5 output now more strictly adheres to the TOA5 file format, but is slower as a result.
 * camp2ascii now returns an Iterator over either a list of files or pandas dataframes instead of just a list of files.
+* Errors caused by corrupt files are now handled by printing an error message and skipping the file, rather than crashing.
 
 # 1.0.2
 * Fixed a bug where the progress bar was miscounting the number of bytes processed.